refactor: server folder re-org
