+++ server/src/filters/pokemon/functions.js
