+++ server/src/graphql/resolvers.js
@@ -6,16 +6,16 @@ const { S2LatLng, S2RegionCoverer, S2LatLngRect } = require('nodes2ts')
 const config = require('@rm/config')
 const { missing, readAndParseJson } = require('@rm/locales')
 
-const buildDefaultFilters = require('../services/filters/builder/base')
-const filterComponents = require('../services/functions/filterComponents')
-const validateSelectedWebhook = require('../services/functions/validateSelectedWebhook')
-const PoracleAPI = require('../services/api/Poracle')
+const buildDefaultFilters = require('../filters/builder/base')
+const filterComponents = require('../utils/filterComponents')
+const validateSelectedWebhook = require('../utils/validateSelectedWebhook')
+const PoracleAPI = require('../services/Poracle')
 const { geocoder } = require('../services/geocoder')
-const scannerApi = require('../services/api/scannerApi')
-const getPolyVector = require('../services/functions/getPolyVector')
-const getPlacementCells = require('../services/functions/getPlacementCells')
-const getTypeCells = require('../services/functions/getTypeCells')
-const { getValidCoords } = require('../services/functions/getValidCoords')
+const scannerApi = require('../services/scannerApi')
+const getPolyVector = require('../utils/getPolyVector')
+const getPlacementCells = require('../utils/getPlacementCells')
+const getTypeCells = require('../utils/getTypeCells')
+const { getValidCoords } = require('../utils/getValidCoords')
 
 /** @type {import("@apollo/server").ApolloServerOptions<import("@rm/types").GqlContext>['resolvers']} */
 const resolvers = {
