+++ server/src/models/Pokestop.js
@@ -4,8 +4,8 @@ const i18next = require('i18next')
 const config = require('@rm/config')
 
 const state = require('../services/state')
-const getAreaSql = require('../services/functions/getAreaSql')
-const { getUserMidnight } = require('../services/functions/getClientTime')
+const getAreaSql = require('../utils/getAreaSql')
+const { getUserMidnight } = require('../utils/getClientTime')
 
 const questProps = {
   quest_type: true,
