+++ server/src/models/Spawnpoint.js
@@ -3,7 +3,7 @@ const { Model, raw } = require('objection')
 const config = require('@rm/config')
 
 const { log, TAGS } = require('@rm/logger')
-const getAreaSql = require('../services/functions/getAreaSql')
+const getAreaSql = require('../utils/getAreaSql')
 
 class Spawnpoint extends Model {
   static get tableName() {
