+++ server/src/routes/rootRouter.js
@@ -10,8 +10,8 @@ const { authRouter } = require('./authRouter')
 const clientRouter = require('./clientRouter')
 const apiRouter = require('./api')
 const state = require('../services/state')
-const areaPerms = require('../services/functions/areaPerms')
-const getServerSettings = require('../services/functions/getServerSettings')
+const areaPerms = require('../utils/areaPerms')
+const getServerSettings = require('../utils/getServerSettings')
 const { secretMiddleware } = require('../middleware/secret')
 const { version } = require('../../../package.json')
