+++ server/src/ui/advMenus.js
@@ -1,5 +1,5 @@
 // @ts-check
-const state = require('../state')
+const state = require('../services/state')
 
 const CATEGORIES = /** @type {const} */ ({
   gyms: ['teams', 'eggs', 'raids', 'pokemon'],
