+++ server/src/utils/consolidateAreas.js
