+++ server/src/utils/getTypeCells.js
