+++ server/src/utils/scannerPerms.js
